# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Lumina HQ API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /rfq/{id}/extract-products:
    post:
      operationId: extractProducts
      tags: [ai]
      summary: AI-extract product list from email body or attachments
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Extracted products
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductsResponse"
  /rfq/{id}/parse-supplier-reply:
    post:
      operationId: parseSupplierReply
      tags: [ai]
      summary: AI-parse pasted supplier reply email to pre-fill quote form
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ParseSupplierReplyInput"
      responses:
        "200":
          description: Parsed quote data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ParsedSupplierReplyResponse"
  /rfq/{id}/draft-supplier:
    post:
      operationId: draftSupplierEmail
      tags: [ai]
      summary: AI-generate supplier inquiry email draft
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Draft content
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DraftResponse"
  /rfq/{id}/draft-customer:
    post:
      operationId: draftCustomerQuote
      tags: [ai]
      summary: AI-generate customer quotation email draft
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerQuoteParams"
      responses:
        "200":
          description: Draft content
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DraftResponse"
  /rfq/{id}/draft-followup:
    post:
      operationId: draftFollowup
      tags: [ai]
      summary: AI-generate follow-up email draft
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Draft content
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FollowupDraftResponse"
  /rfq/{id}/compare:
    get:
      operationId: compareSupplierQuotes
      tags: [ai]
      summary: Get supplier comparison table with landed cost and AI recommendation
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Comparison data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ComparisonResponse"
  /ai/drafts/{draftId}/copied:
    patch:
      operationId: markDraftCopied
      tags: [ai]
      summary: Mark a draft as copied to clipboard
      parameters:
        - in: path
          name: draftId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
...
components:
  schemas:
...
    OkResponse:
      type: object
      required: [ok]
      properties:
        ok:
          type: boolean
...
    RfqProduct:
      type: object
      required: [id, rfqId, productName, attachmentType, extractionConfidence]
      properties:
        id:
          type: integer
        rfqId:
          type: integer
        productName:
          type: string
        catalogueNumber:
          type: ["string", "null"]
        brand:
          type: ["string", "null"]
        quantity:
          type: ["string", "null"]
        specifications:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        attachmentType:
          type: string
        extractionConfidence:
          type: string
...
    SupplierQuoteLine:
      type: object
      required: [id, quoteId, unitPrice, currency]
      properties:
        id:
          type: integer
        quoteId:
          type: integer
        rfqProductId:
          type: ["integer", "null"]
        productName:
          type: ["string", "null"]
        unitPrice:
          type: string
        currency:
          type: string
        leadTimeDays:
          type: ["integer", "null"]
        moq:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
    SupplierQuote:
      type: object
      required: [id, rfqId, supplierName, quotedAt, lines]
      properties:
        id:
          type: integer
        rfqId:
          type: integer
        supplierName:
          type: string
        supplierEmail:
          type: ["string", "null"]
        totalAmount:
          type: ["string", "null"]
        currency:
          type: string
        notes:
          type: ["string", "null"]
        quotedAt:
          type: string
        lines:
          type: array
          items:
            $ref: "#/components/schemas/SupplierQuoteLine"
...
    ProductsResponse:
      type: object
      required: [products]
      properties:
        products:
          type: array
          items:
            $ref: "#/components/schemas/RfqProduct"
        model:
          type: string
    DraftResponse:
      type: object
      required: [draft]
      properties:
        draft:
          type: string
        draftId:
          type: ["integer", "null"]
        model:
          type: string
    FollowupDraftResponse:
      type: object
      required: [draft, daysSinceStageUpdate]
      properties:
        draft:
          type: string
        draftId:
          type: ["integer", "null"]
        model:
          type: string
        daysSinceStageUpdate:
          type: integer
    CustomerQuoteParams:
      type: object
      properties:
        markupPercent:
          type: number
        landedCostBufferPercent:
          type: number
        revisionReason:
          type: string
        parentQuoteId:
          type: integer
    ParseSupplierReplyInput:
      type: object
      required: [emailText]
      properties:
        emailText:
          type: string
        supplierName:
          type: string
    ParsedSupplierReplyResponse:
      type: object
      required: [parsed]
      properties:
        parsed:
          $ref: "#/components/schemas/ParsedQuoteData"
        draftId:
          type: ["integer", "null"]
        model:
          type: string
    ParsedQuoteData:
      type: object
      properties:
        supplierName:
          type: ["string", "null"]
        supplierEmail:
          type: ["string", "null"]
        currency:
          type: string
        notes:
          type: ["string", "null"]
        lines:
          type: array
          items:
            $ref: "#/components/schemas/ParsedQuoteLine"
    ParsedQuoteLine:
      type: object
      required: [productName, unitPrice]
      properties:
        productName:
          type: string
        unitPrice:
          type: string
        currency:
          type: string
        leadTimeDays:
          type: ["integer", "null"]
        moq:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
...
    SupplierPriceCell:
      type: object
      required: [supplier]
      properties:
        supplier:
          type: string
        unitPrice:
          type: ["string", "null"]
        landedUnitPrice:
          type: ["string", "null"]
        sellingUnitPrice:
          type: ["string", "null"]
        currency:
          type: string
        leadTimeDays:
          type: ["integer", "null"]
        moq:
          type: ["string", "null"]
    ComparisonRow:
      type: object
      required: [product, supplierPrices]
      properties:
        product:
          type: string
        catalogueNumber:
          type: ["string", "null"]
        supplierPrices:
          type: array
          items:
            $ref: "#/components/schemas/SupplierPriceCell"
    ComparisonResponse:
      type: object
      properties:
        comparison:
          type: array
          items:
            $ref: "#/components/schemas/ComparisonRow"
        recommendation:
          type: ["string", "null"]
        quotes:
          type: array
          items:
            $ref: "#/components/schemas/SupplierQuote"
        model:
          type: string
        landedCostBufferPercent:
          type: number