# lib/api-zod/src/generated/types/parseSupplierReplyInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Lumina HQ API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ParseSupplierReplyInput {
  emailText: string;
  supplierName?: string;
}