# lib/api-zod/src/generated/types/rfqProduct.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Lumina HQ API specification
 * OpenAPI spec version: 0.1.0
 */

export interface RfqProduct {
  id: number;
  rfqId: number;
  productName: string;
  /** @nullable */
  catalogueNumber?: string | null;
  /** @nullable */
  brand?: string | null;
  /** @nullable */
  quantity?: string | null;
  /** @nullable */
  specifications?: string | null;
  /** @nullable */
  notes?: string | null;
  attachmentType: string;
  extractionConfidence: string;
}